Update engine editor UI

+++ OpenGL-Engine/src/Graphics/Sprite.h
@@ -26,7 +26,7 @@ public:
 
 	bool Sprite_Overlap(Sprite &sprite1, Sprite &sprite2);
 
-	inline const void setPosition(int x, int y) { position.x = x, position.y = y; }
+	inline const void setPosition(float x, float y) { position.x = x, position.y = y; }
 	inline const glm::vec2 getPosition() { return  position; }
 	inline const glm::vec2 getScale() { return  scale; }
 	inline const glm::vec2 getUV() { return  glm::vec2(indexX, indexY); }
